--- src/flag_gems/experimental/__init__.py
@@ -1,5 +1,0 @@
-from .fft import rad2_fft
-
-__all__ = [
-    "rad2_fft",
-]
