+++ tests/test_special_ops.py
@@ -651,3 +651,14 @@ def test_accuracy_cat(shape, dim, dtype):
     with flag_gems.use_gems():
         res_out = torch.cat(inp, dim)
     gems_assert_equal(res_out, ref_out)
+
+
+@pytest.mark.fft
+@pytest.mark.parametrize("shape", [(2048,), (4096,)])
+@pytest.mark.parametrize("dtype", [torch.cfloat])
+def test_accuracy_fft(shape, dtype):
+    x = torch.randn(size=shape, dtype=dtype, device="cuda")
+    y_ref = torch.fft.fft(x)
+    with flag_gems.use_gems():
+        y_res = torch.fft.fft(x)
+    gems_assert_close(y_ref, y_res, dtype)
